requirements.txt
discord.py[voice] == 0.16.1
youtube_dl >= 2017.2.24.1
